# dennic_api_gateway/api/docs/swagger.yaml
definitions:
...
  model_booking_service.CreateDoctorTimeReq:
    properties:
      department_id:
        type: string
      doctor_date:
        type: string
      doctor_id:
        type: string
      end_time:
        type: string
      start_time:
        type: string
      status:
        type: string
    type: object
...
  model_booking_service.DoctorTime:
    properties:
      created_at:
        type: string
      department_id:
        type: string
      doctor_date:
        type: string
      doctor_id:
        type: string
      end_time:
        type: string
      id:
        type: integer
      start_time:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  model_booking_service.DoctorTimesType:
    properties:
      count:
        type: integer
      doctor_times:
        items:
          $ref: '#/definitions/model_booking_service.DoctorTime'
        type: array
    type: object
...
  model_booking_service.UpdateDoctorTimeReq:
    properties:
      department_id:
        type: string
      doctor_date:
        type: string
      doctor_id:
        type: string
      doctor_time_id:
        type: string
      end_time:
        type: string
      start_time:
        type: string
      status:
        type: string
    type: object
...
  model_common.ResponseError:
    properties:
      data:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  model_common.StandardErrorModel:
    properties:
      error:
        $ref: '#/definitions/model_common.ResponseError'
    type: object
...
  models.StatusRes:
    properties:
      status:
        type: boolean
    type: object
host: dennic.uz:9050
info:
  contact: {}
  title: Dennic Project
  version: "1.7"
paths:
...
  /v1/doctor-time:
    delete:
      consumes:
      - application/json
      description: DeleteDoctorTimes - Api for delete Doctor time
      parameters:
      - in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StatusRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: DeleteDoctorTimes
      tags:
      - Doctor Time
    get:
      consumes:
      - application/json
      description: ListDoctorTimes - Api for list doctor time
      parameters:
      - description: searchField
        enum:
        - status
        in: query
        name: searchField
        type: string
      - in: query
        name: limit
        type: string
      - in: query
        name: order_by
        type: string
      - in: query
        name: page
        type: string
      - in: query
        name: value
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_booking_service.DoctorTimesType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: ListDoctorTimes
      tags:
      - Doctor Time
    post:
      consumes:
      - application/json
      description: CreateDoctorTimes - Api for crete doctor time
      parameters:
      - description: CreateDoctorTimeReq
        in: body
        name: CreateDoctorTimeReq
        required: true
        schema:
          $ref: '#/definitions/model_booking_service.CreateDoctorTimeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_booking_service.DoctorTime'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: CreateDoctorTimes
      tags:
      - Doctor Time
    put:
      consumes:
      - application/json
      description: UpdateDoctorTimes - Api for update doctor time
      parameters:
      - description: UpdateDoctorTimeReq
        in: body
        name: UpdateDoctorTimeReq
        required: true
        schema:
          $ref: '#/definitions/model_booking_service.UpdateDoctorTimeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_booking_service.DoctorTime'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: UpdateDoctorTimes
      tags:
      - Doctor Time
  /v1/doctor-time/get:
    get:
      consumes:
      - application/json
      description: GetDoctorTimes - Api for get doctor time
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_booking_service.DoctorTime'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: GetDoctorTimes
      tags:
      - Doctor Time
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"